pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "django-response-formatter"
version = "0.1.0"
description = "Standardize all Django REST Framework API responses into a consistent, predictable format"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Pawan Acharya", email = "zqerpa@gmail.com"},
]
keywords = [
    "django",
    "drf",
    "django-rest-framework",
    "api",
    "response",
    "formatter",
    "standardize",
    "middleware",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "django>=4.2",
    "djangorestframework>=3.14",
]

[project.urls]
Homepage = "https://github.com/suomynonAnonymous/django-response-formatter"
Documentation = "https://github.com/suomynonAnonymous/django-response-formatter#readme"
Repository = "https://github.com/suomynonAnonymous/django-response-formatter"
Issues = "https://github.com/suomynonAnonymous/django-response-formatter/issues"
Changelog = "https://github.com/suomynonAnonymous/django-response-formatter/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-django>=4.5",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
    "mypy>=1.0",
    "django-stubs>=4.0",
]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["django_response_formatter"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["django_response_formatter"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 80
